Correct the spelling of covariants
